+++ server/src/main/kotlin/com/adobe/testing/s3mock/dto/Part.kt
@@ -34,6 +34,16 @@ class Part(
   val lastModified: Date,
   @param:JsonProperty("Size", namespace = "http://s3.amazonaws.com/doc/2006-03-01/")
   val size: Long,
+  @param:JsonProperty("ChecksumCRC32", namespace = "http://s3.amazonaws.com/doc/2006-03-01/")
+  val checksumCRC32: String? = null,
+  @param:JsonProperty("ChecksumCRC32C", namespace = "http://s3.amazonaws.com/doc/2006-03-01/")
+  val checksumCRC32C: String? = null,
+  @param:JsonProperty("ChecksumCRC64NVME", namespace = "http://s3.amazonaws.com/doc/2006-03-01/")
+  val checksumCRC64NVME: String? = null,
+  @param:JsonProperty("ChecksumSHA1", namespace = "http://s3.amazonaws.com/doc/2006-03-01/")
+  val checksumSHA1: String? = null,
+  @param:JsonProperty("ChecksumSHA256", namespace = "http://s3.amazonaws.com/doc/2006-03-01/")
+  val checksumSHA256: String? = null,
 ) {
   constructor(partNumber: Int, etag: String?, size: Long) :
     this(partNumber, normalizeEtag(etag), Date(), size)
@@ -57,6 +67,11 @@ class Part(
     if (size != other.size) return false
     if (lastModified != other.lastModified) return false
     if (etag != other.etag) return false
+    if (checksumCRC32 != other.checksumCRC32) return false
+    if (checksumCRC32C != other.checksumCRC32C) return false
+    if (checksumCRC64NVME != other.checksumCRC64NVME) return false
+    if (checksumSHA1 != other.checksumSHA1) return false
+    if (checksumSHA256 != other.checksumSHA256) return false
 
     return true
   }
@@ -66,6 +81,11 @@ class Part(
     result = 31 * result + size.hashCode()
     result = 31 * result + lastModified.hashCode()
     result = 31 * result + (etag?.hashCode() ?: 0)
+    result = 31 * result + (checksumCRC32?.hashCode() ?: 0)
+    result = 31 * result + (checksumCRC32C?.hashCode() ?: 0)
+    result = 31 * result + (checksumCRC64NVME?.hashCode() ?: 0)
+    result = 31 * result + (checksumSHA1?.hashCode() ?: 0)
+    result = 31 * result + (checksumSHA256?.hashCode() ?: 0)
     return result
   }
 }